Переполнение из за int в perfomance тесте

+++ tasks/romanov_m_closest_elem_vec/tests/performance/main.cpp
@@ -15,7 +15,7 @@ class RomanovMClosestElemVecRunPerfTestProcesses : public ppc::util::BaseRunPerf
     input_data_.resize(vector_size);
 
     for (size_t i = 0; i < vector_size; ++i) {
-      input_data_[i] = (static_cast<int>(i) * 1234567 + 7) % 100000;
+      input_data_[i] = (static_cast<long long>(i) * 1234567 + 7) % 100000;
     }
   }
   // test 2
